+++ tests/charging-station/ocpp/1.6/OCPP16IncomingRequestService-TriggerMessage.test.ts
@@ -184,7 +184,6 @@ await describe('OCPP16IncomingRequestService — TriggerMessage', async () => {
     })
 
     afterEach(() => {
-      mock.reset()
       standardCleanup()
     })
