+++ tests/charging-station/ocpp/2.0/OCPP20IncomingRequestService-RequestStopTransaction.test.ts
@@ -31,41 +31,16 @@ import {
   OCPP20ReasonEnumType,
 } from '../../../../src/types/ocpp/2.0/Transaction.js'
 import { Constants } from '../../../../src/utils/index.js'
-import { standardCleanup } from '../../../../tests/helpers/TestLifecycleHelpers.js'
+import { flushMicrotasks, standardCleanup } from '../../../../tests/helpers/TestLifecycleHelpers.js'
 import { TEST_CHARGING_STATION_BASE_NAME } from '../../ChargingStationTestConstants.js'
 import { createMockChargingStation } from '../../ChargingStationTestUtils.js'
 import { createMockAuthService } from '../auth/helpers/MockFactories.js'
 import {
+  createOCPP20ListenerStation,
   resetConnectorTransactionState,
   resetLimits,
   resetReportingValueSize,
 } from './OCPP20TestUtils.js'
-
-/**
- * @param baseName - Base name for the mock charging station
- * @returns The mock station and its request handler spy
- */
-function createListenerStation (baseName: string): {
-  requestHandlerMock: ReturnType<typeof mock.fn>
-  station: ChargingStation
-} {
-  const requestHandlerMock = mock.fn(async () => Promise.resolve({}))
-  const { station } = createMockChargingStation({
-    baseName,
-    connectorsCount: 3,
-    evseConfiguration: { evsesCount: 3 },
-    heartbeatInterval: Constants.DEFAULT_HEARTBEAT_INTERVAL,
-    ocppRequestService: {
-      requestHandler: requestHandlerMock,
-    },
-    stationInfo: {
-      ocppStrictCompliance: false,
-      ocppVersion: OCPPVersion.VERSION_201,
-    },
-    websocketPingInterval: Constants.DEFAULT_WEBSOCKET_PING_INTERVAL,
-  })
-  return { requestHandlerMock, station }
-}
 
 await describe('F03 - Remote Stop Transaction', async () => {
   let mockStation: ChargingStation
@@ -73,7 +48,7 @@ await describe('F03 - Remote Stop Transaction', async () => {
   let testableService: ReturnType<typeof createTestableIncomingRequestService>
 
   beforeEach(() => {
-    const { station } = createListenerStation(TEST_CHARGING_STATION_BASE_NAME)
+    const { station } = createOCPP20ListenerStation(TEST_CHARGING_STATION_BASE_NAME)
     mockStation = station
     incomingRequestService = new OCPP20IncomingRequestService()
     testableService = createTestableIncomingRequestService(incomingRequestService)
@@ -246,7 +221,7 @@ await describe('F03 - Remote Stop Transaction', async () => {
     let listenerStation: ChargingStation
 
     beforeEach(() => {
-      ;({ requestHandlerMock, station: listenerStation } = createListenerStation(
+      ;({ requestHandlerMock, station: listenerStation } = createOCPP20ListenerStation(
         TEST_CHARGING_STATION_BASE_NAME + '-LISTENER'
       ))
       listenerService = new OCPP20IncomingRequestService()
@@ -356,7 +331,7 @@ await describe('F03 - Remote Stop Transaction', async () => {
       )
 
       // Flush microtask queue so .catch(errorHandler) executes
-      await Promise.resolve()
+      await flushMicrotasks()
 
       assert.strictEqual(transactionEventCallCount, 1)
     })
